changes about contact us with API

+++ UI/JS/dash.js
@@ -1,82 +1,7 @@
-//Blog menu
-// const blogImage = document.getElementById("blogImage");
-// let imageUrl;
-// blogImage.addEventListener("change", function () {
-//   const fileReader = new FileReader();
-//   fileReader.addEventListener("load", () => {
-//     imageUrl = fileReader.result;
-//   });
-//   fileReader.readAsDataURL(this.files[0]);
-// });
-
-// let blogCards = document.getElementById("blog-cards");
-// document.getElementById("blog-form").addEventListener("submit", function (e) {
-//   e.preventDefault();
-//   let title = document.getElementById("blogTitle");
-//   let description = document.getElementById("blogMessage");
-//   let blogSubmit = document.getElementById("blogSubmit");
-
-//   let myHeaders = new Headers();
-//   myHeaders.append("Content-Type", "application/json");
-
-//   let data = {
-//     image: imageUrl,
-//     title: title.value,
-//     description: description.value,
-//   };
-
-//   let requestOptions = {
-//     method: "POST",
-//     headers: myHeaders,
-//     body: JSON.stringify(data),
-//     redirect: "follow",
-//   };
-
-//   fetch("http://localhost:5000/api/blogs", requestOptions)
-//     .then((response) => response.json())
-//     .then((result) => console.log(result))
-//     .catch((error) => console.log("error", error));
-//   blogSubmit.textContent = "Blog added well";
-//   document.getElementById("blog-form").reset();
-// });
-
-// //Get Blog
-// document.addEventListener("DOMContentLoaded", function () {
-//   var requestOptions = {
-//     method: "GET",
-//     redirect: "follow",
-//   };
-//   fetch("http://localhost:5000/api/blogs")
-//     .then((response) => response.json())
-//     .then((data) => {
-//       console.log(data);
-//       for (let blog of data) {
-//         let card = document.createElement("div");
-//         card.classList.add("card");
-//         let image = document.createElement("img");
-//         image.src = blog.image;
-//         card.appendChild(image);
-
-//         let title = document.createElement("h2");
-//         title.textContent = blog.title;
-//         card.appendChild(title);
-
-//         let description = document.createElement("p");
-//         description.textContent = blog.description;
-//         card.appendChild(description);
-
-//         blogCards.appendChild(card);
-//       }
-//     })
-//     .catch((error) => console.log("error", error));
-// });
-
 const MAX_WIDTH = 300;
 const MAX_HEIGHT = 300;
-
 const blogImage = document.getElementById("blogImage");
 let imageUrl;
-
 blogImage.addEventListener("change", function () {
   const fileReader = new FileReader();
   fileReader.addEventListener("load", () => {
@@ -86,7 +11,6 @@ blogImage.addEventListener("change", function () {
   });
   fileReader.readAsDataURL(this.files[0]);
 });
-
 function resizeImage(imageUrl, maxWidth, maxHeight, callback) {
   const img = new Image();
   img.src = imageUrl;
@@ -109,7 +33,6 @@ function resizeImage(imageUrl, maxWidth, maxHeight, callback) {
     const canvas = document.createElement("canvas");
     canvas.width = width;
     canvas.height = height;
-
     const ctx = canvas.getContext("2d");
     ctx.drawImage(img, 0, 0, width, height);
     const resizedImageUrl = canvas.toDataURL("image/jpeg", 0.7);
@@ -117,13 +40,11 @@ function resizeImage(imageUrl, maxWidth, maxHeight, callback) {
   };
 }
 
+let title = document.getElementById("blogTitle");
+let description = document.getElementById("blogMessage");
 let blogCards = document.getElementById("blog-cards");
-
 document.getElementById("blog-form").addEventListener("submit", function (e) {
   e.preventDefault();
-
-  let title = document.getElementById("blogTitle");
-  let description = document.getElementById("blogMessage");
   let blogSubmit = document.getElementById("blogSubmit");
 
   let myHeaders = new Headers();
@@ -151,17 +72,21 @@ document.getElementById("blog-form").addEventListener("submit", function (e) {
 });
 
 //Get Blog
+
+let card;
 document.addEventListener("DOMContentLoaded", function () {
   var requestOptions = {
     method: "GET",
     redirect: "follow",
   };
-  fetch("http://localhost:5000/api/blogs")
+
+  fetch("http://localhost:5000/api/blogs", requestOptions)
     .then((response) => response.json())
     .then((data) => {
       console.log(data);
       for (let blog of data) {
-        let card = document.createElement("div");
+        card = document.createElement("div");
+        card.classList.add("blog-cards");
         card.classList.add("card");
         let image = document.createElement("img");
         image.src = blog.image;
@@ -169,70 +94,54 @@ document.addEventListener("DOMContentLoaded", function () {
 
         let title = document.createElement("h2");
         title.textContent = blog.title;
+        title.classList.add("blog-list-title");
         card.appendChild(title);
 
         let description = document.createElement("p");
-        description.textContent = blog.descripti;
+        description.textContent = blog.description;
         card.appendChild(description);
 
-        // Create an "Update" button for each blog post
-        let updateBtn = document.createElement("button");
-        updateBtn.textContent = "Update";
-        updateBtn.addEventListener("click", () => {
-          // Handle the click event for the "Update" button
-          console.log(`Update button clicked for blog post ID ${blog._id}`);
-        });
-        card.appendChild(updateBtn);
+        let btnContainer = document.createElement("div");
+        btnContainer.classList.add("btnContainer");
 
+        let deleteBtn = document.createElement("button");
+        deleteBtn.textContent = "Delete";
+        deleteBtn.classList.add("deleteBtn");
+        deleteBtn.addEventListener("click", () => {
+          if (confirm("Are you sure you want to delete this blog post?")) {
+            let requestOptions = {
+              method: "DELETE",
+              redirect: "follow",
+            };
+
+            fetch(`http://localhost:5000/api/blogs/${blog._id}`, requestOptions)
+              .then((response) => {
+                if (response.status === 204) {
+                  console.log(blog);
+                  location.reload();
+                }
+              })
+              .catch((error) => console.log("error", error));
+          }
+        });
+
+        let updateBtn = document.createElement("button");
+
+        updateBtn.textContent = "Update";
+        updateBtn.classList.add("updateBtn");
+        updateBtn.addEventListener("click", () => {
+          console.log(blog._id);
+        });
+
+        btnContainer.appendChild(deleteBtn);
+        btnContainer.appendChild(updateBtn);
+
+        card.appendChild(btnContainer);
         blogCards.appendChild(card);
       }
     })
     .catch((error) => console.log("error", error));
 });
-
-// const blogs = blogMessages
-//   .map((item) => {
-//     const blog = `
-//     <div class="blog-card" id="blog-card">
-//   <img src="${item.image}" alt="" class="imgPreview" />
-//    <h3 class="blog-list-title" id="list-heading">${item.title}</h3>
-//     <p class="blog-list-description" id="content">${item.message}</p>
-//     <div class="del-edit">
-//     <button onclick="deleteBlog(${item.index})" class="blog-delete">Delete</button>
-//     <button class="edit-btn">Edit</button>
-//     </div>
-//     </div>
-//     `;
-//     return blog;
-//   })
-//   .join("");
-
-// window.addEventListener("load", function () {
-//   blogCards.innerHTML = blogs;
-// });
-
-//Blog Delete
-
-// function deleteBlog(index) {
-//   blogMessages = blogMessages.filter((blog) => blog.index !== index);
-//   localStorage.setItem("blogInfo", JSON.stringify(blogMessages));
-//   blogCards.innerHTML = blogMessages
-//     .map((item) => {
-//       const blog = `
-//       <div class="blog-card" id="blog-card">
-//         <img src="${item.image}" alt="" class="imgPreview" />
-//         <h3 class="blog-list-title" id="list-heading">${item.title}</h3>
-//         <p class="blog-list-description" id="content">${item.message}</p>
-//         <div class="del-edit">
-//           <button onclick="deleteMessage(${item.index})" class="blog-delete">Delete</button>
-//           <button class="edit-btn">Edit</button>
-//         </div>
-//       </div>
-//       `;
-//       return blog;
-//     })
-//     .join("");
-// }
 
 //Contact
 const url = "http://localhost:5000/api/contacts";
@@ -298,6 +207,11 @@ fetch(url)
   });
 
 const deleteContact = async (id) => {
+  const confirmed = confirm("Are you sure you want to delete this contact?");
+  if (!confirmed) {
+    return;
+  }
+
   const getToken = JSON.parse(localStorage.getItem("usertoken"));
   const settings = {
     method: "DELETE",
